+++ NEWS.md
@@ -1,5 +1,9 @@
 # qol 1.3.5 - DEVELOPMENT
 
+### Changed functionality
+
+* `summarise_plus()`: When only passing `class` variables without `values` and using the default `statistics` then only frequencies will be calculated instead of sums and frequencies, which would be identical. (23.08.2026)
+
 ### Fixed
 
 * `combine_into_workbook()`: Fixed table of contents example not working as intended. (21.08.2026)